✨ Search icon, smartphone navigation and article card title underline
Added the search icon in the navigation bar, it will now be displayed on the right at all time, the menu on the left and the logo in the middle while the tablet navigation will be logo on the left, menu and search icon on the right.

The article card's title now has a thin underline.